requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
attrs==25.3.0
Automat==25.4.16
bcrypt==4.3.0
black==25.1.0
certifi==2025.8.3
click==8.2.1
constantly==23.10.4
decorator==5.2.1
dnspython==2.7.0
email-validator==2.3.0
fastapi==0.116.1
fastapi-cli==0.0.10
fastapi-cloud-cli==0.1.5
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy_extensions==1.1.0
outcome==1.3.0.post0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.8.0
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-asyncio-cooperative==0.40.0
pytest-sugar==1.1.1
pytest-tornasync==0.6.0.post2
pytest-trio==0.8.0
pytest-twisted==1.14.3
python-decouple==3.8
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
sentry-sdk==2.35.2
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.43
starlette==0.47.3
termcolor==3.1.0
tornado==6.5.2
trio==0.30.0
Twisted==25.5.0
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
zope.interface==7.2
